LayoutTests/fast/dom/HTMLElement/iscontenteditable-designmodeon-allinherit-subframe.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<div id="result"></div>
<script>
description('document.designMode inherits from parent documents.')

function runTest()
{
    evalAndLog('document.designMode = "on"');

    subDocument = document.getElementById("subframe").contentDocument;
    shouldBe('subDocument.getElementById("div").contentEditable', '"inherit"');
    shouldBe('subDocument.getElementById("div").isContentEditable', 'true');

    document.getElementById("subframe").style.display= 'none';
}
</script>
<script src="../../js/resources/js-test-post.js"></script>
<iframe id="subframe" onload="runTest()" src="data:text/html,<div id='div'></div>"></iframe>
</body>
</html>
