LayoutTests/fast/dom/HTMLElement/iscontenteditable-designmodeon-allinherit-subframe-expected.txt

document.designMode inherits from parent documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
document.designMode = "on"
PASS subDocument.getElementById("div").contentEditable is "inherit"
PASS subDocument.getElementById("div").isContentEditable is true
